fix: Ensure `todaySlots` is given to today date in delivery times
The bug is in generating the delivery time dates and slots, when today
is not available by the composite delivery rule for the order it is
filtered, then the logic assign `todaySlots` - which is filtered according
to `same_day_deadline` - to the first valid date which is not today.

The solution is to check that the first date - in the generated dates
is today before assigning `todaySlots` to the first valid date.